% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{KNN_Julia}
\alias{KNN_Julia}
\title{KNN_Julia}
\usage{
KNN_Julia(
  sommet,
  saison,
  nationalite,
  role,
  annee,
  sexe,
  age,
  pro,
  solo,
  oxygene,
  mort = 0,
  blesse = 0,
  nvoisin = 5
)
}
\arguments{
\item{sommet}{nom du sommet}

\item{saison}{choix de la saison}

\item{nationalite}{quelle nationalité ?}

\item{role}{rôle dans la cordée}

\item{annee}{année de l'ascension}

\item{sexe}{votre genre}

\item{age}{votre âge}

\item{pro}{êtes vous un sherpa}

\item{solo}{comptez vous faire l'ascension en solo}

\item{oxygene}{utilsation d'oxygène}

\item{mort}{êtes vous prêt à mourir}

\item{blesse}{avez vous peur de vous blesser}

\item{nvoisin}{nombre de voisins}
}
\description{
KNN_Julia
}
\examples{
KNN_Julia("Ama Dablam","Autumn","France","Climber",2025,1,20,0,0,0,0,0,5)
}
