% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{RandomForest_Julia}
\alias{RandomForest_Julia}
\title{RandomForest_Julia}
\usage{
RandomForest_Julia(
  sommet,
  saison,
  nationalite,
  role,
  annee,
  sexe,
  age,
  pro,
  solo,
  oxygene,
  mort = 0,
  blesse = 0,
  ntree = 5
)
}
\arguments{
\item{sommet}{nom du sommet}

\item{saison}{choix de la saison}

\item{nationalite}{quelle nationalité ?}

\item{role}{rôle dans la cordée}

\item{annee}{année de l'ascension}

\item{sexe}{votre genre}

\item{age}{votre âge}

\item{pro}{êtes vous un sherpa}

\item{solo}{comptez vous faire l'ascension en solo}

\item{oxygene}{utilsation d'oxygène}

\item{mort}{êtes vous prêt à mourir}

\item{blesse}{avez vous peur de vous blesser}

\item{ntree}{nombre d'arbres dans la RandomForest}
}
\description{
RandomForest_Julia
}
\examples{
RandomForest_Julia("Ama Dablam","Autumn","France","Climber",2025,1,20,0,0,0,0,0,5)
}
